doc2txt_frontend/pyproject.toml
[project]
name = "doc2txt_frontend"
version = "0.1"
dependencies = [
   "PyJWT==2.8.0",
   "requests==2.32.3",
   "streamlit==1.36.0"
]